Revert "Unpack double-packed gz"
This reverts commit ba582042840893a49122413d731b8869c066d97b.

+++ build.sh
@@ -5,13 +5,8 @@ function version() {
 }
 
 function build() {
-  rm -rf BUILD BUILDROOT RPMS SRPMS || true
-  test -s SOURCES/atlassian-$1-$(version $1).tar.gz
-  if [[ $? -ne 0 ]]; then
-    wget https://downloads.atlassian.com/software/${2:-$1}/downloads/atlassian-${3:-$1}-$(version $1).tar.gz -O SOURCES/atlassian-$1-$(version $1).tar.gz
-    gunzip SOURCES/atlassian-$1-$(version $1).tar.gz
-    mv SOURCES/atlassian-$1-$(version $1).tar SOURCES/atlassian-$1-$(version $1).tar.gz
-  fi
+	rm -rf BUILD BUILDROOT RPMS SRPMS || true
+  test -s SOURCES/atlassian-$1-$(version $1).tar.gz || wget https://downloads.atlassian.com/software/${2:-$1}/downloads/atlassian-${3:-$1}-$(version $1).tar.gz -O SOURCES/atlassian-$1-$(version $1).tar.gz
   rpmbuild --define "_topdir $(git rev-parse --show-toplevel)" -bb SPECS/$1.spec
   mv RPMS/x86_64/$1-*.rpm .
 }
